+++ src/eventEmitter.ts
@@ -1,12 +1,12 @@
 import { EventEmitter } from 'events';
-import { statsBar } from './statsBar';
+import { customBar } from './customBar';
 import { command } from './command';
 
-class ExtensionEmitter extends EventEmitter {}
+class ExtensionEmitter extends EventEmitter { }
 
 export const extensionEmitter = new ExtensionEmitter();
 
 extensionEmitter.on('setting-update', () => {
-  statsBar.onSettingUpdate();
+  customBar.onSettingUpdate();
   command.changeContext();
 });