+++ src/sysinfo/index.ts
@@ -2,6 +2,12 @@ import * as si from 'systeminformation';
 import * as os from 'os';
 import { getMacOsMemoryUsageInfo } from './memory';
 import { isDarwin, isWin32 } from '../utils';
+import { exec } from 'child_process';
+import { promisify } from 'util';
+
+const execPromise = promisify(exec);
+import { spawn } from 'child_process';
+
 
 export function siInit() {
   if (isWin32) {
@@ -15,32 +21,19 @@ export function siRelease() {
   }
 }
 
-export async function getCpuSpeed() {
-  try {
-    const res = await si.cpuCurrentSpeed();
-    return res.avg;
-  } catch (err) {}
-}
-
 export async function getCpuLoad() {
   try {
-    const res = await si.currentLoad();
-    return res.currentLoad;
-  } catch (err) {}
-}
-
-export async function getLoadavg() {
-  try {
-    const res = os.loadavg();
-    return res;
-  } catch (err) {}
-}
-
-export async function getIP() {
-  const defaultInterface = await si.networkInterfaceDefault();
-  const res = await si.networkInterfaces();
-  const cur = res.find(item => item.iface === defaultInterface);
-  return cur?.ip4;
+    const valueObject = {
+      currentLoad: 'currentLoad',
+      mem: 'total,used'
+    }
+    const data = await si.get(valueObject);
+    return {
+      cpu_used: data.currentLoad.currentLoad,
+      total: data.mem.total,
+      used: data.mem.used,
+    };
+  } catch (err) { }
 }
 
 export async function getNetworkSpeed() {
@@ -52,13 +45,7 @@ export async function getNetworkSpeed() {
       up: cur.tx_sec,
       down: cur.rx_sec
     };
-  } catch (err) {}
-}
-
-export async function getUpTime() {
-  try {
-    return os.uptime();
-  } catch (err) {}
+  } catch (err) { }
 }
 
 export async function getMemoryUsage() {
@@ -80,15 +67,87 @@ export async function getMemoryUsage() {
         active: res.active
       };
     }
-  } catch (err) {}
+  } catch (err) { }
+}
+
+async function getX86GpuInfo() {
+  try {
+    const { stdout, stderr } = await execPromise('nvidia-smi --query-gpu=memory.total,memory.used,utilization.gpu --format=csv,noheader,nounits');
+    if (stderr) {
+      return null;
+    }
+
+    const lines = stdout.trim().split('\n');
+    const gpuInfo = lines.map(line => {
+      const [totalMemory, usedMemory, gpuUtilization] = line.split(', ').map(Number);
+      return {
+        totalMemory,
+        usedMemory,
+        gpuUtilization,
+      };
+    });
+    return gpuInfo;
+  } catch (error) {
+    return null;
+  }
+}
+
+async function getTegraGpuInfo() {
+  try {
+    const { stdout, stderr } = await execPromise('tegrastats');
+    if (stderr) {
+      return null;
+    }
+
+    const lines = stdout.trim().split('\n');
+    const tegraInfo = lines.map(line => {
+      const match = line.match(/(\d+)\/(\d+)\s+MB\s+(\d+)%/);
+      if (match) {
+        const usedMemory = Number(match[1]);
+        const totalMemory = Number(match[2]);
+        const gpuUtilization = Number(match[3]);
+        return {
+          usedMemory,
+          totalMemory,
+          gpuUtilization,
+        };
+      }
+      return null;
+    }).filter(Boolean);
+
+    return tegraInfo;
+  } catch (error) {
+    return null;
+  }
+}
+
+async function GetGpuInfo() {
+  try {
+    const { stdout } = await execPromise('uname -m');
+    const arch = stdout.trim();
+
+    if (arch === 'x86_64' || arch === 'i686') {
+      return await getX86GpuInfo();
+    } else if (arch.includes('tegra')) {
+      return await getTegraGpuInfo();
+    } else {
+      return null;
+    }
+  } catch (error) {
+    return null;
+  }
+}
+
+export async function getGpuLoad() {
+  try {
+    return GetGpuInfo();
+  } catch (err) { }
 }
 
 export const sysinfoData = {
-  cpuLoad: getCpuLoad,
-  loadavg: getLoadavg,
   networkSpeed: getNetworkSpeed,
-  memoUsage: getMemoryUsage,
-  uptime: getUpTime
+  cpuLoad: getCpuLoad,
+  gpuLoad: getGpuLoad,
 };
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
@@ -101,8 +160,6 @@ export type StatsModule = keyof typeof sysinfoData;
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const StatsModuleNameMap: { [key in StatsModule]: string } = {
   cpuLoad: 'CpuLoad',
-  loadavg: 'Loadavg',
   networkSpeed: 'NetworkSpeed',
-  memoUsage: 'MemoryUsage',
-  uptime: 'Uptime'
+  gpuLoad: 'GpuLoad'
 };